Refactor imports in GitProjectSyncDropdown and GitSyncDropdown for consistency
