+++ packages/insomnia/src/ui/components/dropdowns/git-sync-dropdown.tsx
@@ -20,9 +20,9 @@ import { useGitProjectPushActionFetcher } from '~/routes/git.push';
 import { useGitProjectRepoFetcher } from '~/routes/git.repo';
 import { useGitProjectResetActionFetcher } from '~/routes/git.reset';
 import { useGitProjectStatusActionFetcher } from '~/routes/git.status';
+import { getOauth2FormatName } from '~/sync/git/get-oauth2-format-name';
 
 import type { GitRepository } from '../../../models/git-repository';
-import { getOauth2FormatName } from '../../../sync/git/utils';
 import type { MergeConflict } from '../../../sync/types';
 import { ConfigLink } from '../github-app-config-link';
 import { Icon } from '../icon';
